Got rid of StoreID and StoreName duality.
The store names are gone for good, and the
new StoreIDs are structural (i.e. they can
be nested) and plain text. They are
human-readable and act both as a unique
identifier as well as a relative file path
to the store directory. We still call them
"StoreId"s and not "StoreName"s because
renaming a store is no longer supported.