Working on debugging the off-by-one error in the files.
The problem seems to be that the block header
contains a command count which is 1 higher than
it's supposed to be. Parsing always fails at the
last index, and the previous command always
consumed exactly the remainder of the buffer. So
either the last command was counted but not
written, or the index was somewhere increased
where it should not have been. It's weird.