Removed versioning features, simplified API accordingly.
It turns out that having version control baked into the
storage engine (even as an optional feature) is not very
smart, because:

- If you have versioning, you need to take care of
  rollovers and the like to make it scalable.

- Rollovers need to be transparent to the user.

- To make this happen, the management of secondary
  indices would also have to happen at this level.

- This means that we would have to be able to
  extract individual property values from an
  arbitrary byte array.

We've been there before with ChronoDB, and we
definitely DON'T want to go there again. Whatever
abstraction layer manages the secondary indices
needs to be able to enforce a data schema, and
that's not the key-value store for sure.