Compaction no longer drops all historic entries.
Instead, it drops only entries which have newer versions, and the older
versions are inaccessible by any currently ongoing transaction. This allows
us to compact files even while a transaction is currently reading them, as
the new (compacted) file will still contain the same snapshot data as the
original file (from the perspective of the transaction).