Separated Cursor<K,V> and CursorInternal<K,V> [WIP]
The goal here is to get rid of modcounts altogether. A cursor can either be a top-level cursor (i.e. it has no parent) or a child cursor (i.e. it has a parent). If a cursor has a parent, all its public API is "shut off" and will throw exceptions when called.

The CursorInternal<K,V> interface provides an internal API (which mirrors the public one) which is still eligible for being called even if the cursor has a parent. This internal API is designed to be used only by the parent cursor.

This allows us to cleanly separate owned cursors from top-level cursors without keeping track of modcounts. This also allows us to remove levels and levels of abstract base classes which reduce performance for cursors and make profiling very hard due to deep call stacks.